requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
boto3==1.28.17
botocore==1.31.17
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Fiona==1.9.4.post1
fonttools==4.41.1
frozenlist==1.4.0
geopandas==0.13.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
jmespath==1.0.1
kiwisolver==1.4.4
libretranslatepy==2.1.1
load-dotenv==0.1.0
lxml==4.9.3
matplotlib==3.7.2
multidict==6.0.4
numpy==1.25.2
openai==0.27.8
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
shapely==2.0.1
six==1.16.0
tqdm==4.65.0
translate==3.6.1
tzdata==2023.3
urllib3==1.26.16
yarl==1.9.2
zipp==3.16.2
